+++ packages/orq-rc/src/orq_ai_sdk/models/get_v2_human_eval_sets_id_op.py
@@ -83,7 +83,7 @@ class GetV2HumanEvalSetsIDResponseBody2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-02T09:20:24.795Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-02T10:46:54.630Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -168,7 +168,7 @@ class GetV2HumanEvalSetsIDResponseBody1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-02T09:20:24.795Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-02T10:46:54.630Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")