+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN6QXCKNYWTZWKZ5E655BNGD"
+        "01KN6WVS8DZB3J9CKYAB6M38F4"
     )
     r"""The unique identifier of the data source"""