+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -1903,7 +1903,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KN6QXC5H8P6EFEVT7BGW31TW"
+    id: Optional[str] = "01KN6WVRSVRZ85WHNMKEBTMP49"
 
     description: Optional[str] = None