+++ packages/orq-rc/RELEASES.md
@@ -6819,3 +6819,13 @@ Based on:
 - [python v4.7.0-rc.42] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.42 - packages/orq-rc
+
+## 2026-04-02 10:49:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.43] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.43 - packages/orq-rc